Clean up and defensive programming
