+++ gherkin/util/__init__.py
@@ -1,4 +1,4 @@
-from gherkin.util.utilities import generate_random_goal
+from gherkin.util.utilities import generate_random_goal, generate_visualizer
 from gherkin.util.visualizer import Visualizer
 
-__all__ = ["generate_random_goal", "Visualizer"]
+__all__ = ["generate_random_goal", "generate_visualizer", "Visualizer"]
