+++ gherkin/util/utilities.py
@@ -1,6 +1,7 @@
 import numpy as np
 
-from gherkin.model import Goal, Angle
+from gherkin.model import Goal, Angle, World
+from gherkin.util.visualizer import Visualizer
 
 
 def generate_random_goal(min_radius: float, max_radius: float) -> Goal:
@@ -16,3 +17,13 @@ def generate_random_goal(min_radius: float, max_radius: float) -> Goal:
     y = int(r * np.sin(theta))
     angle = Angle(np.random.randint(low=0, high=180))
     return Goal(x, y, angle)
+
+
+def generate_visualizer(num_robots: int):
+    height = 300
+    robot_width = 300
+    total_width = robot_width * num_robots
+    robot_origins = [(int(robot_width / 2) + (robot_width * i), int(height / 2)) for i in range(num_robots)]
+    world = World(total_width, robot_width, height, robot_origins)
+    vis = Visualizer(world)
+    return vis
